supercan/d5035: switch to two independenty devices
CAN0 + CAN1 used to share a command pipe. This is no longer the case.
In fact, each CAN now has its own interface with one cmd pipe and one msg pipe.

The rational behind this is that on Windows we get two different device files which can be operated in parallel.