superdfu/can: fix flashing with dfu-util from Windows 10
The reason why most open source USB projects request an extra PID
is likely that they need to assign the WinUSB v6.1 driver to the
interface under Windows 10. With a single PID, Windows gets confused
whether the device is in Runtime or DFU mode.

To actually have the dfu-util pick up the device on Windows 10,
the WinUSB v6.1 driver needs to be assigned to the SuperDFU interface.
Note this applies to both Runtime _and_ DFU mode. Zadig is a handy
tool to perform this assignment.

dfu-util 0.9 doesn't appear to respect overriding vendor/product id
from the command line. This can result in the device being stuck in
DFU mode (PID=0x5036). In this mode, I couldn't get dfu-util for flash
a file target at PID 0x5035. The solution adopted here is to _not_
set vendor/product id on the firmware file but to leave them at 0xffff
which seems to work fine.

Another issue with dfu-util under Windows 10 appears to be its
inability to reset the USB after having requested the device to detach.
To work around this issue, whenever SuperCAN receives DFU_REQUEST_DETACH
it will simply start the bootloader. Note that this behaviour violates
the DFU 1.1 specification.