supercan: add true nop message for CAN epp
SC_MSG_CAN_NOP is 4 bytes of filler. This message is now used to fill
up an URB to the transport size.

This is done in an attempt to work around an issue on Windows in which
it is sometimes not possible to determine the order of the returned URB
buffers.