SuperCAN: initial version with 1MHz TC0/1 counter
The RX/TX timestamps are now taken from the counter.
The counter value register is updated with upon request (hw cmd) on
the next tick of the clock, that is 1 [us] later at the earliest.

Busy waiting for the counter in the interrupt handler doesn't appear
to work right. Therefore the current counter is requested early on
in the interrupt handler and then blindly used once it is required.
This obviously introduces a race condition between the tick frequency
of the counter and the execution time of the code path taken through
the ISR. While this is bad, I can't currently think of another way
of accessing the counter fast enough.