SuperCAN: ensure CAN TSW is running
Without TSW, a prolong period of inactivity on the CAN will
result in no interrupts being fired and hence no USB traffic.
This results in the host missing TC0/1 counter laps.

Enabling the M_CAN TSW counter (without using its value) provides
a constant stream of irqs which is sufficient in to track time
on the host.