SuperCAN: wait for 1MHz counter in CAN interrupt handler
The interrupt handler now always waits for the counter value.
This way a highly recent timestamp [us] will be used to compute
RX/TX times.

Unlike previous iterations of the code, this version of the wait
appears to work. Tested with the harness.

Non interrupt code now atomically requests a counter update to
prevent a race condition with the interrupt service routine.