SuperCAN: don't reset timestamp counter on bus on/off
This change is to support a multi client scenario.

Initially, all clients are on the bus. Then, the controlling client
goes off bus. At a later time it goes back on the bus. Previously the
reset timestamp would cause the other clients, that are still active
but don't receive any message b/c the device is off the bus, to then
receive messages from the past (until the duration the device was
previously on the bus has been reached again).

With this change, the clients oblivious to the fact that the device
went off the bus, will continue to receive timestamps once the device
is back on the bus as if it never went offline.