D5035-01: fix bad use of PLLs
According to Microchips documentation the PLL should output at
least 96 MHz which it didn't previously (80/48 MHz). This is now
fixed. As a side effect, PLL1 gets freed up as the the USB now
uses DFLL48M. The FLL is configured to run in closed loop from the
external 16 MHz crystal.