SuperDFU: initial support for working bootloader update
The update takes advantage of SAME5X dual bank flash memory
which can map either bank to 0x00000000.

The bootloader uses the same DFU app header as the application does.
When writing the bootloader to flash, the app header is replaced by
the vector table following the application header aligned to 1K.