SuperCAN: SAM E5x: attempt to reset CAN on start
When M_CAN was in error-passive prior to the device being taken
offline by the driver, the error counters used to be retained
when the device would come back online. This would result in the device
reporting error-passive from the get go which is not what one expects.

This change detecs this particular condition befor the devices comes
online for the driver and forces the transmission/reception of CAN
messages until the error counters no longer change. This typically (?)
decrements the counters enough that error-passive goes away.