SuperDFU: add support for auto setup on Windows
There are three parts to this story:
- enabling MS OS 2.0 descriptors
- get the device to be recognized as composite device
- ensure dfu-util is able to claim the DFU interface

The first part requires the .bcdUSB to be 0x210 or better and
to have the necessary callbacks

	tud_descriptor_bos_cb
	tud_vendor_control_xfer_cb

in place. tud_vendor_control_xfer_cb needs to be setup
to handle the request for the MS OS 2.0 descriptor data.

For the second part to work, the device needs to be configured
as composite through the descriptor.

The third part requires some sort of device GUID to be attached.
Without the GUID, dfu-util doesn't seem to be able to claim the device.