D5035-50: add doc comment for custom timer reset code
Microchip SAM E51 timerm (TC) reset doesn't appear to work as
advertised in DS60001507E-page 1717. The work around seems to
be to only start the timer when it is actually stopped.

Note there is a race condition around checking the
flag and and starting the timer. However, this should only
be a problem if we are very close to expiration, which we prevent
by resetting the COUNT register prior to the check.