slLIN: D5035-50: fixes sercom getting out of sync
There was a feedback loop due to the unit being in auto baud
mode. Errors during bit timing of the sync field would accumulate
to the point where the baud register was totally off the target
baud rate.

To address the issues, reset the baud register prior to each
transmission of a header.