slLIN: Trinket M0: improve BREAK detection
Previous detection flagged break after having received 0x00 with form error set.
This change  a timer at the end of reveived zero byte to wait for the remaining
three bit times (13 bit times minimal). When the timer elapses, the rx pin is
sampled for zero. If the pin is in fact zero, BREAK is detected.