SuperCAN: stm32h7a: various changes to track down timestamp issues
For both rx and tx paths sometimes timestamps are jumpbled. At appears
this is related to the reconstruction of the timestamps for Bosch
M_CAN based chips.

What happens is that for a batch of rx or tx frames the timestamp
of frames towards the begin of the batch (received / transmitted
earlier) could be cmputed in such a way that they appeaer before
the last timestamp of the previous batch.

While this shouldn't obviously be possible, an easy solution would be
to mark every frame with the same timestamp.